krgrpc: Introduce communicators and connections
An rpc_connection encapsulates the communication state with another Kerrighed
node for a given Kerrighed instance on both nodes. When nodes join and leave a
cluster, a node must establish (resp. close) connections with all other nodes
of the cluster.

The communication state of a node with all nodes of the cluster is encaspsulated
in a rpc_communicator. An rpc_communicator contains pointers to all current
connections with the other nodes of the cluster, plus more global state like the
next id to use for a new RPC.

When a nodes joins the cluster, all nodes add their connection with the joining
node to their communicator, and so does the joining node with its connections
with the other nodes. When a node leaves the cluster, its connections are
removed from the communicators of all nodes before being closed.

A krg_namespace owns a single rpc_communicator, which id is the same on all
nodes of the cluster. This rpc_communicator is created when the new namespace is
created, and destroyed with the namespace too. A kddm namespace owns the same
rpc_communicator as the krg_namespace based on it.

These notions of communicator and connection will help making Kerrighed's
hotplug add and remove reliable with respect to the communication layer.

This commit introduces the notion of rpc_communicator and rpc_connection, but
without all the lifecycle things yet. A static rpc_communicator and boot-time-
initialized rpc_connections replace current global variables that will be
dynamically allocated when the whole lifecycle stuff will be implemented.

Next commits progressively transform krgrpc to make rpc_communicator and
rpc_connection truely dynamic.

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>