krgsched: Using Notifier API instead of module_hook_*
This patch introduce the use of notifier API instead of the old
module_hook_* scheme.

It changes the call to modifier_hook_* functions to the corresponding
ones of Notifier API.

It removes both module_hook.h and module_hook.c which are useless now
(sorry). Also updates the Makefile according.

The old struct module_hook_desc declarations are replaced by
ATOMIC_NOTIFIER_HEAD().

The upstream probes that were using module_hooks are updated to the new
API (that is, migration_probe and mosix_probe).

Code is running fine since a couple of weeks now.

Signed-off-by: Alexandre Lissy <alexandre.lissy@etu.univ-tours.fr>
Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>