hotplug: Introduce the hotplug coordinator (core)
This commit introduces the hotplug coordinator, a mechanism that will be
used to serialize add and remove operations, and thus will make concurrent
nodes_add() and nodes_remove() safe.

The coordinator is hosted on the node having lowest id, and moves during
add and remove operations to remain on the node having lowest id.

To serialize hotplug operations, each operation must be first queued in a
workqueue using hotplug_queue_request(). This workqueue serializes hotplug
operations locally. hotplug_queue_request() fails if the node is being
removed or not yet added to a cluster.

When the work is executed, it must serialize with hotplug operations from
other nodes by calling hotplug_start_request() before starting the actual
add/remove operation. hotplug_start_request() notifies the coordinator and
waits for its acknowledgement. hotplug_start_request() fails if the node
is being removed, or if the coordinator could not be contacted.

Once the hotplug operation is finished, the work must notify the
coordinator that the operation is done, by calling hotplug_finish_request().
Failing to contact the coordinator leads to blocking all future hotplug
operations.

Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>