krgsched: Introducing the Mattload probe.
Mattload Probe is an implementation of the load measure described in
http://portail.scd.univ-tours.fr/search*frf/X?PEROTIN,%20MATTHIEU&m=t&m=u
as part of Matthieu Pérotin's PhD thesis.

It makes uses of all the previously introduced probe, providing a way to
compute a Unix-like load but that also considers the presence of a local
user on the box.

This new probe exposes two probe sources:
 * mattload, returning the load value computed as described in the PhD
   thesis linked above.
 * load_increment, which returns the value of load that a new process
   will have on this specific node.

This new load is computed using a couple of informations:
 - length of the runqueue
 - presence of an ``interactive user''
and some parameters to ponderate the previous informations. This allows
us to limit the scheduling toward one node that is used by an
``interactive user'' (e.g. a student using the computer).

Concretely, (1 + K*(state))((runqueue length)/speed).

[ Removed abusively restricting dependencies on other probes in Kconfig
  -- Louis Rilling ]

Signed-off-by: Alexandre Lissy <alexandre.lissy@etu.univ-tours.fr>
Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>