krgsched: Introducing Local User Presence tools.
This patch introduce several tools aimed at notifying the kerrighed
kernel of the presence of an "interactive user".

It consists of two parts:
 * a kernel module implementing a simple counter with a kernel API:
   the local_user_presence module.
 * a kernel module providing a Kerrighed SchedConfig probe, called
   user_probe.

The kernel module counter exposes a userspace /proc filesystem, allowing
to :
 * Notify of connection/disconnection of a user
 * Get the number of connected used
 * Check if the node is free
 * Check if the node is used

The Kerrighed probe currently provides only one probe source: connected,
which just returns the number of connected users to the node retrieved
from the local_user_presence module.

This is part of the API written by Mathieu Dabert. It will be used by
the mattload probe which will be introduced later.

[ Re-ordered menuconfig entries in a more natural way -- Louis Rilling ]

Signed-off-by: Alexandre Lissy <alexandre.lissy@etu.univ-tours.fr>
Signed-off-by: Louis Rilling <louis.rilling@kerlabs.com>