[BUGFIX] Fix possible double put on KDDM objects and rare issues in kddm_remove_frozen.
The optimization introduced in generic_kddm_grab_object and
generic_kddm_get_object to avoid the useless creation of objects in
the no_ft case can lead to 2 issues.

1st issue: lets assume the following scenario:
   kddm_grab_object_no_ft
   _kddm_remove_frozen_object

If the object does not exist and the local node is the default owner,
the object will not be created and thus, will not be really grabbed.
The remove_frozen code will then BUG since we are removeing a non
grabbed object.

2nd issue: lets assume the following scenario:
       Process P1                   Process P2
  kddm_grab_object_no_ft
                                  kddm_grab_object
  kddm_put_object
                                  kddm_put_object

If the object does not exist and the local node is the default owner,
the object will not be created in process P1 and thus, will not be
really grabbed.
If the process P2 grab the object before process P1 does its put, the
object will be created and the put in P1 will do a real put. We will
then do 2 real puts (one in P1 and one in P2) since only one real grab
has been done.

This patch fixes these issues by removeing the object allocation
optimization. Any call to get or grab now creates an object, with no
exception.